requirements.txt
SQLAlchemy~=1.4.43
psycopg2~=2.9.5
python-dotenv~=0.21.0